+++ cards-array.js
@@ -1,0 +1,75 @@
+const tuzPik = `<img src="./static/img/tuz-pik.png" alt="card">`
+const kingPik = `<img src="./static/img/king-pik.png" alt="card">`
+const ladyPik = `<img src="./static/img/lady-pik.png" alt="card">`
+const valPik = `<img src="./static/img/val-pik.png" alt="card">`
+const tenPik = `<img src="./static/img/10-pik.png" alt="card">`
+const ninePik = `<img src="./static/img/9-pik.png" alt="card">`
+const eightPik = `<img src="./static/img/8-pik.png" alt="card">`
+const sevenPik = `<img src="./static/img/7-pik.png" alt="card">`
+const sixPik = `<img src="./static/img/6-pik.png" alt="card">`
+const tuzCherv = `<img src="./static/img/a-hrt.png" alt="card">`
+const kingCherv = `<img src="./static/img/king-hrt.png" alt="card">`
+const ladyCherv = `<img src="./static/img/lady-hrt.png" alt="card">`
+const valCherv = `<img src="./static/img/val-hrt.png" alt="card">`
+const tenCherv = `<img src="./static/img/10-hrt.png" alt="card">`
+const nineCherv = `<img src="./static/img/9-hrt.png" alt="card">`
+const eightCherv = `<img src="./static/img/8-hrt.png" alt="card">`
+const sevenCherv = `<img src="./static/img/7-hrt.png" alt="card">`
+const sixCherv = `<img src="./static/img/6-hrt.png" alt="card">`
+const tuzBubn = `<img src="./static/img/a-b.png" alt="card">`
+const kingBubn = `<img src="./static/img/k-b.png" alt="card">`
+const ladyBubn = `<img src="./static/img/l-b.png" alt="card">`
+const valBubn = `<img src="./static/img/v-b.png" alt="card">`
+const tenBubn = `<img src="./static/img/10-b.png" alt="card">`
+const nineBubn = `<img src="./static/img/9-b.png" alt="card">`
+const eightBubn = `<img src="./static/img/8-b.png" alt="card">`
+const sevenBubn = `<img src="./static/img/7-b.png" alt="card">`
+const sixBubn = `<img src="./static/img/6-b.png" alt="card">`
+const tuzKrest = `<img src="./static/img/a-k.png" alt="card">`
+const kingKrest = `<img src="./static/img/k-k.png" alt="card">`
+const ladyKrest = `<img src="./static/img/l-k.png" alt="card">`
+const valKrest = `<img src="./static/img/v-k.png" alt="card">`
+const tenKrest = `<img src="./static/img/10k.png" alt="card">`
+const nineKrest = `<img src="./static/img/9-k.png" alt="card">`
+const eightKrest = `<img src="./static/img/8-k.png" alt="card">`
+const sevenKrest = `<img src="./static/img/7-k.png" alt="card">`
+const sixKrest = `<img src="./static/img/6-k.png" alt="card">`
+
+export let cardsArray = [
+    tuzPik,
+    kingPik,
+    ladyPik,
+    valPik,
+    tenPik,
+    ninePik,
+    eightPik,
+    sevenPik,
+    sixPik,
+    tuzCherv,
+    kingCherv,
+    ladyCherv,
+    valCherv,
+    tenCherv,
+    nineCherv,
+    eightCherv,
+    sevenCherv,
+    sixCherv,
+    tuzBubn,
+    kingBubn,
+    ladyBubn,
+    valBubn,
+    tenBubn,
+    nineBubn,
+    eightBubn,
+    sevenBubn,
+    sixBubn,
+    tuzKrest,
+    kingKrest,
+    ladyKrest,
+    valKrest,
+    tenKrest,
+    nineKrest,
+    eightKrest,
+    sevenKrest,
+    sixKrest,
+]
