+++ index.js
@@ -1,7 +1,12 @@
 import './style.css'
+import {
+    renderEasyGame,
+    renderMediumGame,
+    renderHardGame,
+} from './render-game.js'
 
 let appHtml = null
-let appEl = document.getElementById('app')
+export let appEl = document.getElementById('app')
 //check
 function renderMainPage() {
     appHtml = `<div class="flex-container">
@@ -32,383 +37,6 @@ function renderMainPage() {
     appEl.innerHTML = appHtml
 }
 renderMainPage()
-
-const tuzPik = `<img src="./static/img/tuz-pik.png" alt="card">`
-const kingPik = `<img src="./static/img/king-pik.png" alt="card">`
-const ladyPik = `<img src="./static/img/lady-pik.png" alt="card">`
-const valPik = `<img src="./static/img/val-pik.png" alt="card">`
-const tenPik = `<img src="./static/img/10-pik.png" alt="card">`
-const ninePik = `<img src="./static/img/9-pik.png" alt="card">`
-const eightPik = `<img src="./static/img/8-pik.png" alt="card">`
-const sevenPik = `<img src="./static/img/7-pik.png" alt="card">`
-const sixPik = `<img src="./static/img/6-pik.png" alt="card">`
-const tuzCherv = `<img src="./static/img/a-hrt.png" alt="card">`
-const kingCherv = `<img src="./static/img/king-hrt.png" alt="card">`
-const ladyCherv = `<img src="./static/img/lady-hrt.png" alt="card">`
-const valCherv = `<img src="./static/img/val-hrt.png" alt="card">`
-const tenCherv = `<img src="./static/img/10-hrt.png" alt="card">`
-const nineCherv = `<img src="./static/img/9-hrt.png" alt="card">`
-const eightCherv = `<img src="./static/img/8-hrt.png" alt="card">`
-const sevenCherv = `<img src="./static/img/7-hrt.png" alt="card">`
-const sixCherv = `<img src="./static/img/6-hrt.png" alt="card">`
-const tuzBubn = `<img src="./static/img/a-b.png" alt="card">`
-const kingBubn = `<img src="./static/img/k-b.png" alt="card">`
-const ladyBubn = `<img src="./static/img/l-b.png" alt="card">`
-const valBubn = `<img src="./static/img/v-b.png" alt="card">`
-const tenBubn = `<img src="./static/img/10-b.png" alt="card">`
-const nineBubn = `<img src="./static/img/9-b.png" alt="card">`
-const eightBubn = `<img src="./static/img/8-b.png" alt="card">`
-const sevenBubn = `<img src="./static/img/7-b.png" alt="card">`
-const sixBubn = `<img src="./static/img/6-b.png" alt="card">`
-const tuzKrest = `<img src="./static/img/a-k.png" alt="card">`
-const kingKrest = `<img src="./static/img/k-k.png" alt="card">`
-const ladyKrest = `<img src="./static/img/l-k.png" alt="card">`
-const valKrest = `<img src="./static/img/v-k.png" alt="card">`
-const tenKrest = `<img src="./static/img/10k.png" alt="card">`
-const nineKrest = `<img src="./static/img/9-k.png" alt="card">`
-const eightKrest = `<img src="./static/img/8-k.png" alt="card">`
-const sevenKrest = `<img src="./static/img/7-k.png" alt="card">`
-const sixKrest = `<img src="./static/img/6-k.png" alt="card">`
-
-let cardsArray = [
-    tuzPik,
-    kingPik,
-    ladyPik,
-    valPik,
-    tenPik,
-    ninePik,
-    eightPik,
-    sevenPik,
-    sixPik,
-    tuzCherv,
-    kingCherv,
-    ladyCherv,
-    valCherv,
-    tenCherv,
-    nineCherv,
-    eightCherv,
-    sevenCherv,
-    sixCherv,
-    tuzBubn,
-    kingBubn,
-    ladyBubn,
-    valBubn,
-    tenBubn,
-    nineBubn,
-    eightBubn,
-    sevenBubn,
-    sixBubn,
-    tuzKrest,
-    kingKrest,
-    ladyKrest,
-    valKrest,
-    tenKrest,
-    nineKrest,
-    eightKrest,
-    sevenKrest,
-    sixKrest,
-]
-
-function getRandomInt(cardNumber) {
-    let minValue = 0
-    let maxValue = 35
-    const randomCards = []
-    for (let i = 0; i < cardNumber; i++) {
-        randomCards[i] =
-            Math.round(Math.random() * (maxValue - minValue)) + minValue
-    }
-    return [...randomCards, ...randomCards]
-    // return Math.round(Math.random() * (maxValue - minValue)) + minValue
-}
-
-function shuffle(array) {
-    for (let i = array.length - 1; i > 0; i--) {
-        const j = Math.floor(Math.random() * (i + 1))
-
-        ;[array[i], array[j]] = [array[j], array[i]]
-    }
-    return array
-}
-
-function renderEasyGame() {
-    appHtml = `<div class="container">
-    <header class="header">
-        <div class="header__timer">
-<div class="header__text-box">
-<p class="header__text">min</p>
-<p class="header__text">sec</p>
-</div>
-<div class="header__count">
-00.00
-</div>
-        </div>
-        <button class="header__button button">
-            Начать заново
-        </button>
-    </header>
-    <div class="cards-field">
-    </div>
-</div>`
-
-    appEl.innerHTML = appHtml
-
-    const randomCardsIds = getRandomInt(3)
-
-    const shuffledArray = shuffle(randomCardsIds)
-
-    const cardField = document.querySelector('.cards-field')
-
-    const cardElements = []
-    for (let i = 0; i < shuffledArray.length; i++) {
-        var element = shuffledArray[i]
-        var cardElement = `<div class="cards-hide"><div data-id="${element}" class="cards-field__card" id="card-vision">${cardsArray[element]}</div></div>`
-        cardElements.push(cardElement)
-    }
-
-    cardField.innerHTML = cardElements.join('')
-
-    let cardsVision = document.querySelectorAll('.cards-hide')
-    for (const cardVision of cardsVision) {
-        setTimeout(() => {
-            cardVision.classList.add('picture')
-        }, 5000)
-        cardVision.addEventListener('click', () => {
-            cardVision.classList.remove('picture')
-        })
-    }
-    let clickCards = document.querySelectorAll('.cards-field__card')
-    let openCards = []
-    for (const clickCard of clickCards) {
-        setTimeout(() => {
-            clickCard.classList.add('opacity')
-        }, 5000)
-
-        clickCard.addEventListener('click', () => {
-            clickCard.classList.remove('opacity')
-            let cardValue = clickCard.dataset.id
-
-            openCards.push(cardValue)
-            console.log(openCards)
-            if (openCards.length == 2) {
-                if (openCards[0] != openCards[1]) {
-                    alert('вы проиграли!')
-                }
-            }
-            if (openCards.length == 4) {
-                if (openCards[2] != openCards[3]) {
-                    alert('вы проиграли!')
-                }
-            }
-            if (openCards.length == 6) {
-                setTimeout(() => {
-                    alert('вы выиграли!')
-                }, 1000)
-            }
-        })
-    }
-}
-function renderMediumGame() {
-    appHtml = `<div class="container">
-    <header class="header">
-        <div class="header__timer">
-<div class="header__text-box">
-<p class="header__text">min</p>
-<p class="header__text">sec</p>
-</div>
-<div class="header__count">
-00.00
-</div>
-        </div>
-        <button class="header__button button">
-            Начать заново
-        </button>
-    </header>
-    <div class="cards-field">
-    </div>
-</div>`
-
-    appEl.innerHTML = appHtml
-
-    const randomCardsIds = getRandomInt(6)
-
-    const shuffledArray = shuffle(randomCardsIds)
-
-    const cardField = document.querySelector('.cards-field')
-
-    const cardElements = []
-    for (let i = 0; i < shuffledArray.length; i++) {
-        var element = shuffledArray[i]
-        var cardElement = `<div class="cards-hide"><div data-id="${element}" class="cards-field__card" id="card-vision">${cardsArray[element]}</div></div>`
-        cardElements.push(cardElement)
-    }
-
-    cardField.innerHTML = cardElements.join('')
-
-    let cardsVision = document.querySelectorAll('.cards-hide')
-    for (const cardVision of cardsVision) {
-        setTimeout(() => {
-            cardVision.classList.add('picture')
-        }, 5000)
-        cardVision.addEventListener('click', () => {
-            cardVision.classList.remove('picture')
-        })
-    }
-    let clickCards = document.querySelectorAll('.cards-field__card')
-    let openCards = []
-    for (const clickCard of clickCards) {
-        setTimeout(() => {
-            clickCard.classList.add('opacity')
-        }, 5000)
-
-        clickCard.addEventListener('click', () => {
-            clickCard.classList.remove('opacity')
-            let cardValue = clickCard.dataset.id
-
-            openCards.push(cardValue)
-            console.log(openCards)
-            if (openCards.length == 2) {
-                if (openCards[0] != openCards[1]) {
-                    alert('вы проиграли!')
-                }
-            }
-            if (openCards.length == 4) {
-                if (openCards[2] != openCards[3]) {
-                    alert('вы проиграли!')
-                }
-            }
-            if (openCards.length == 6) {
-                if (openCards[4] != openCards[5]) {
-                    alert('вы проиграли!')
-                }
-            }
-            if (openCards.length == 8) {
-                if (openCards[6] != openCards[7]) {
-                    alert('вы проиграли!')
-                }
-            }
-            if (openCards.length == 10) {
-                if (openCards[8] != openCards[9]) {
-                    alert('вы проиграли!')
-                }
-            }
-            if (openCards.length == 12) {
-                setTimeout(() => {
-                    alert('вы выиграли!')
-                }, 1000)
-            }
-        })
-    }
-}
-
-function renderHardGame() {
-    appHtml = `<div class="container">
-    <header class="header">
-        <div class="header__timer">
-<div class="header__text-box">
-<p class="header__text">min</p>
-<p class="header__text">sec</p>
-</div>
-<div class="header__count">
-00.00
-</div>
-        </div>
-        <button class="header__button button">
-            Начать заново
-        </button>
-    </header>
-    <div class="cards-field">
-    </div>
-</div>`
-
-    appEl.innerHTML = appHtml
-
-    const randomCardsIds = getRandomInt(9)
-
-    const shuffledArray = shuffle(randomCardsIds)
-
-    const cardField = document.querySelector('.cards-field')
-
-    const cardElements = []
-    for (let i = 0; i < shuffledArray.length; i++) {
-        var element = shuffledArray[i]
-        var cardElement = `<div class="cards-hide"><div data-id="${element}" class="cards-field__card" id="card-vision">${cardsArray[element]}</div></div>`
-        cardElements.push(cardElement)
-    }
-
-    cardField.innerHTML = cardElements.join('')
-
-    let cardsVision = document.querySelectorAll('.cards-hide')
-    for (const cardVision of cardsVision) {
-        setTimeout(() => {
-            cardVision.classList.add('picture')
-        }, 5000)
-        cardVision.addEventListener('click', () => {
-            cardVision.classList.remove('picture')
-        })
-    }
-    let clickCards = document.querySelectorAll('.cards-field__card')
-    let openCards = []
-    for (const clickCard of clickCards) {
-        setTimeout(() => {
-            clickCard.classList.add('opacity')
-        }, 5000)
-
-        clickCard.addEventListener('click', () => {
-            clickCard.classList.remove('opacity')
-            let cardValue = clickCard.dataset.id
-
-            openCards.push(cardValue)
-            console.log(openCards)
-            if (openCards.length == 2) {
-                if (openCards[0] != openCards[1]) {
-                    alert('вы проиграли!')
-                }
-            }
-            if (openCards.length == 4) {
-                if (openCards[2] != openCards[3]) {
-                    alert('вы проиграли!')
-                }
-            }
-            if (openCards.length == 6) {
-                if (openCards[4] != openCards[5]) {
-                    alert('вы проиграли!')
-                }
-            }
-            if (openCards.length == 8) {
-                if (openCards[6] != openCards[7]) {
-                    alert('вы проиграли!')
-                }
-            }
-            if (openCards.length == 10) {
-                if (openCards[8] != openCards[9]) {
-                    alert('вы проиграли!')
-                }
-            }
-            if (openCards.length == 12) {
-                if (openCards[10] != openCards[11]) {
-                    alert('вы проиграли!')
-                }
-            }
-            if (openCards.length == 14) {
-                if (openCards[12] != openCards[13]) {
-                    alert('вы проиграли!')
-                }
-            }
-            if (openCards.length == 16) {
-                if (openCards[14] != openCards[15]) {
-                    alert('вы проиграли!')
-                }
-            }
-            if (openCards.length == 18) {
-                setTimeout(() => {
-                    alert('вы выиграли!')
-                }, 1000)
-            }
-        })
-    }
-}
-
-
 
 let easyLevelGame = document.getElementById('easy')
 let mediumLevelGame = document.getElementById('medium')
